Update centroid files for AxxSolder v3.2
The centroid file located in the PCBWay community build is for v3.0, and there does not appear to be an updated centroid file for v3.2 that includes the two new resistors (R49 and R50). This caused an error when PCBWay attempted to manufacture a board for me, so I loaded the project in KiCAD and created new .pos files for them. Figure they should be included here for anyone else that wants to get an AxxSolder from PCBWay while AxxAxx is on his world tour.